+++ src/ak_sap/Model/model.py
@@ -3,6 +3,7 @@ import typing
 
 from .constants import _UNITS, _UNITS_LITERALS, _PROJECT_INFO_KEYS
 from ak_sap.utils import log, MasterClass
+from ak_sap.utils.decorators import smooth_sap_do
 
 class Model(MasterClass):
     def __init__(self, mySapObject) -> None:
@@ -30,14 +31,13 @@ class Model(MasterClass):
         log.debug(f'Queried Database Units: {_units}')
         return _units
     
+    @smooth_sap_do
     def set_units(self, value: _UNITS_LITERALS):
         """Updates the current units of model"""
         log.info(f'Updating the current units to {value}')
-        try:
-            _unit_to_set = _UNITS.index(value) + 1
-            assert self.SapModel.SetPresentUnits(_unit_to_set) == 0
-        except Exception as e:
-            log.critical(str(e))
+        _unit_to_set = _UNITS.index(value) + 1
+        return self.SapModel.SetPresentUnits(_unit_to_set)
+
         
     @property
     def merge_tol(self) -> float:
@@ -50,13 +50,11 @@ class Model(MasterClass):
             log.critical(str(e) + f'Return Tol: {tol}')
         return tol
     
+    @smooth_sap_do
     def set_merge_tol(self, value: float):
         """Sets the program auto merge tolerance"""
         log.info(f'Updating the auto-merge tolerance to {value}')
-        try:
-            assert self.SapModel.SetMergeTol(value) == 0
-        except Exception as e:
-            log.critical(str(e))
+        return self.SapModel.SetMergeTol(value)
     
     @property
     def filepath(self) -> Path:
@@ -105,23 +103,20 @@ class Model(MasterClass):
         log.debug('Extracting Sap logs')
         return self.SapModel.GetUserComment()[0]
     
+    @smooth_sap_do
     def set_logs(self, value: str) -> None:
         """sets the user comments and log data."""
-        log.info(f'Adding "{value}" to SAP logs')
-        try:
-            assert self.SapModel.SetUserComment(value) == 0
-        except Exception as e:
-            log.critical(str(e))
-            
+        return self.SapModel.SetUserComment(value)
+
     def lock(self):
         """Lock Model"""
-        self._update_lock(lock=True)
+        self.__update_lock(lock=True)
     
     def unlock(self):
         """Unlock Model"""
-        self._update_lock(lock=False)
+        self.__update_lock(lock=False)
         
-    def _update_lock(self, lock: bool):
+    def __update_lock(self, lock: bool):
         """Lock/Unlock Model"""
         try:
             assert self.SapModel.SetModelIsLocked(lock) == 0
