+++ src/ak_sap/utils/credentials.py
@@ -1,22 +1,12 @@
 import getpass
-from pathlib import Path
-import sys
+import keyring
 
 from . import log
 
-if sys.platform=="win32":
-    import getpass
-    import keyring
-
-
 def getpwd(item, username):
-    if sys.platform=="win32":
-        log.debug(f'getting pwd for `{username}` in `{item}`')
-        pwd = keyring.get_password(item, username)
-        if not pwd:
-            log.warning("Password is not saved in keyring.")
-            pwd = getpass.getpass(f"Enter the password for {item} corresponding to Username:{username}: ")
-    else:
-        pwd = getpass.getpass(f"Enter the password for `{item}` corresponding to Username:`{username}`  : ")
-    
+    log.debug(f'getting pwd for `{username}` in `{item}`')
+    pwd = keyring.get_password(item, username)
+    if not pwd:
+        log.warning("Password is not saved in keyring.")
+        pwd = getpass.getpass(f"Enter the password for {item} corresponding to Username:{username}: ")
     return pwd
